refactor(#3282): move __any_box_*/__any_unbox_* to sibling module any-boxing-helpers.ts
Supersedes the same-file extraction: the LOC-regrowth ratchet flagged that a
same-file helper GROWS the god-file (+25). Per #3282's stated approach, extract
to a new SIBLING MODULE so the god-file SHRINKS. Both the box family (7 helpers)
and the unbox family (4 helpers) move verbatim into
src/codegen/any-boxing-helpers.ts as registerAnyBoxHelpers /
registerAnyUnboxHelpers; ensureAnyHelpers calls them. addHelper threaded in as a
callback → registration order + bodies unchanged.

- prove-emit-identity check: IDENTICAL 56/56 (gc/standalone/wasi/linear).
- any-helpers.ts 2752 → 2425 LOC (−327); LOC-regrowth ratchet green, no allowance.
- tsc 0; biome/prettier/loc-budget clean.
- undefinedSingletonActive back-import is a runtime-only ESM cycle (safe).

✓ opus-1
